Re-factor SGRID variables
